infra/bigquery_export_spark/requirements.txt
google-cloud-bigquery==3.23
google-cloud-storage==2.16
google-cloud-firestore==2.20.1
# pyspark==3.5.5
